+++ modules/canvaskit/externs.js
@@ -109,7 +109,6 @@ var CanvasKit = {
 	},
 
 	ParagraphStyle: function() {},
-
 	RSXFormBuilder: function() {},
 	SkColorBuilder: function() {},
 	SkRectBuilder: function() {},
@@ -229,6 +228,12 @@ var CanvasKit = {
 		_makeShader: function() {},
 	},
 
+	SkImageFilter: {
+		MakeBlur: function() {},
+		MakeColorFilter: function() {},
+		MakeCompose: function() {},
+	},
+
 	SkMatrix: {
 		identity: function() {},
 		invert: function() {},
@@ -238,6 +243,10 @@ var CanvasKit = {
 		scaled: function() {},
 		skewed: function() {},
 		translated: function() {},
+	},
+
+	SkMaskFilter: {
+		MakeBlur: function() {},
 	},
 
 	SkPaint: {
@@ -255,6 +264,7 @@ var CanvasKit = {
 		setBlendMode: function() {},
 		setColor: function() {},
 		setFilterQuality: function() {},
+		setImageFilter: function() {},
 		setMaskFilter: function() {},
 		setPathEffect: function() {},
 		setShader: function() {},
@@ -421,6 +431,9 @@ var CanvasKit = {
 	UnderlineDecoration: {},
 	OverlineDecoration: {},
 	LineThroughDecoration: {},
+
+	SaveLayerInitWithPrevious: {},
+	SaveLayerF16ColorType: {},
 
 	Affinity: {
 		Upstream: {},